use connection manangement and just replace it
